+++ packages/react-day-picker/.vscode/tasks.json
@@ -1,0 +1,13 @@
+{
+  "version": "2.0.0",
+  "tasks": [
+    {
+      "type": "npm",
+      "script": "typecheck",
+      "problemMatcher": ["$tsc-watch"],
+      "label": "npm: typecheck",
+      "detail": "tsc --project ./tsconfig.json --noEmit",
+      "isBackground": true
+    }
+  ]
+}
