+++ packages/react-day-picker/src/types/Labels.ts
@@ -1,8 +1,3 @@
-import { DayLabelFormatter } from './DayLabelFormatter';
-import { NavButtonLabelFormatter } from './NavButtonLabelFormatter';
-import { WeekdayLabelFormatter } from './WeekdayLabelFormatter';
-import { WeekNumberLabelFormatter } from './WeekNumberLabelFormatter';
-
 /**
  * Map of functions to translate ARIA labels for the relative elements.
  */
@@ -15,3 +10,37 @@ export type Labels = {
   labelWeekday: WeekdayLabelFormatter;
   labelWeekNumber: WeekNumberLabelFormatter;
 };
+
+import { Locale } from 'date-fns';
+
+import { ModifierStatus } from '../types';
+
+/**
+ * Represent a function to format the ARIA label for the [[Day]] component.
+ */
+export type DayLabelFormatter = (
+  day: Date,
+  modifiers: ModifierStatus,
+  options?: { locale?: Locale }
+) => string;
+
+/**
+ * Represent a function to format the ARIA label for the "next month" / "prev
+ * month" buttons in the navigation.
+ */
+export type NavButtonLabelFormatter = (
+  month?: Date,
+  options?: { locale?: Locale }
+) => string;
+
+/** Represent a function to format the ARIA label for the Head component.*/
+export type WeekdayLabelFormatter = (
+  day: Date,
+  options?: { locale?: Locale }
+) => string;
+
+/** Represent a function to format the ARIA label of the week number.*/
+export type WeekNumberLabelFormatter = (
+  n: number,
+  options?: { locale?: Locale }
+) => string;
