--- packages/react-day-picker/src/types/WeekdayLabelFormatter.ts
@@ -1,5 +1,0 @@
-/** Represent a function to format the ARIA label for the Head component.*/
-export type WeekdayLabelFormatter = (
-  day: Date,
-  options?: { locale?: Locale }
-) => string;
